[Weather Demo] Re-URI.encode city name
Prior to this commit the weather example fails for cities which have
characters which must be `URI.encode`d in order to be represented in a
URL, e.g. "Düsseldorf", "Kraków", "Malmö", "Tromsø". This commit adds a
call to `URI.encode` before passing the city name along to `wttr.in`.